Fix crash after upgrade due to ambiguous column
Commit a0c8fe8f47d8cb322ae56ab49c22ce6707ec63bc ("Filter SIM contacts with
flag") extended the query to join accounts with the raw_contacts table. Both
tables have an "account_name" column which results in an "ambiguous column name"
error. Fix this by prefixing the column by the accounts (as name_accounts)
table.

This issue is triggered whenever the views are regenerated (for example, when
updating from JB to KK). The user notices that an app (acore?) crashed on boot
and that the dialer is disfunctional.

Change-Id: Ia6c5ea788dbaf9506367933fe7051103c5b30696